# Improve layout/paint performance in editor while running a project