… test/unit/components/__snapshots__/sprite-selector-item.test.jsx.snap