This PR introduces a couple of layout/css tricks in order to vastly improve the layout+paint+composition performance of the editor while a typical projects is running. The main improvement is to apply the same layout trick that is used in the sprite library to allow the sprite selector images to change frequently without causing a layout of the target pane as a whole. That makes it so much less layout/painting has to happen:

![birds-develop](https://user-images.githubusercontent.com/654102/51263957-5345bc00-1983-11e9-8735-2be239ab69c8.gif)
![birds-fixed](https://user-images.githubusercontent.com/654102/51263959-5345bc00-1983-11e9-80d7-ab0988145651.gif)

This improves the composition time by almost a half on both low-power machines (chromebook composition time goes from 20ms to 10ms) and high-power machines (my laptop goes from 5ms to 2.5ms).

I see an almost 50% improvement in frame times on the low-power chromebook using the "Birds" project (id=198711193) which was chosen from the front page.

Prior to this change, the react updates were a smaller part of frame time than the layout+paint+comp part, but this change makes the react portion the dominant blocker for frame time (on this project). 

FYI the changes to the selector/backpack is to make those work with the style changes on the sprite-selector-item, since that component is used in those places as well.